refactor(app): extract the query-tab wiring
New tab, run-into-a-new-result-tab, switching and closing result tabs,
closing a query tab, renaming one, and tab selection move to
`wire::tabs`.

`AppFns` bundles the long-lived closures `main` builds — rebuild the
query tree, load editor text, save/restore a tab, guard pending edits,
run SQL — so a wiring module takes two arguments instead of a growing
list of individual callbacks. Named type aliases keep the struct readable.
