+++ app/src/wire/mod.rs
@@ -8,4 +8,5 @@ pub(crate) mod conn_form;
 pub(crate) mod edit;
 pub(crate) mod grid;
 pub(crate) mod settings;
+pub(crate) mod tabs;
 pub(crate) mod update;
